+++ h3ronpy/src/transform.rs
@@ -14,6 +14,7 @@ pub struct Transform {
 
 #[pymethods]
 impl Transform {
+    #[allow(clippy::many_single_char_names)] // using the same parameter names as the affine library
     #[new]
     pub fn new(a: f64, b: f64, c: f64, d: f64, e: f64, f: f64) -> Self {
         Self {